run apache-rat.
