+++ s2core/src/main/scala/org/apache/s2graph/core/features/S2GraphFeatures.scala
@@ -22,8 +22,9 @@ package org.apache.s2graph.core.features
 import org.apache.tinkerpop.gremlin.structure.Graph.Features
 import org.apache.tinkerpop.gremlin.structure.Graph.Features.GraphFeatures
 
-
 class S2GraphFeatures extends GraphFeatures {
+
+
   override def supportsComputer(): Boolean = false
 
   override def supportsThreadedTransactions(): Boolean = false
